Introduce GetScanJobTolerations on ConfigData
