+++ pkg/vulnerabilityreport/scanner.go
@@ -2,7 +2,6 @@ package vulnerabilityreport
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"github.com/aquasecurity/starboard/pkg/apis/aquasecurity/v1alpha1"
 	"github.com/aquasecurity/starboard/pkg/docker"
@@ -116,24 +115,17 @@ func (s *Scanner) getCredentials(ctx context.Context, spec corev1.PodSpec, ns st
 	return kube.MapContainerNamesToDockerAuths(kube.GetContainerImagesFromPodSpec(spec), imagePullSecrets)
 }
 
-func getScanJobTolerations(starboardConfigData starboard.ConfigData) []corev1.Toleration {
-	scanJobTolerations := []corev1.Toleration{}
-	tolerationsJsonFromConfigMap := starboardConfigData["scanJob.tolerations"]
-	klog.V(3).Infof("scanJob.tolerations: %s", tolerationsJsonFromConfigMap)
-	err := json.Unmarshal([]byte(tolerationsJsonFromConfigMap), &scanJobTolerations)
-	if err != nil {
-		klog.V(3).Infof("error unmarshalling scanJob.tolerations: %s", err)
-	}
-	return scanJobTolerations
-}
-
 func (s *Scanner) prepareScanJob(workload kube.Object, spec corev1.PodSpec, credentials map[string]docker.Auth, starboardConfigData starboard.ConfigData) (*batchv1.Job, []*corev1.Secret, error) {
 	templateSpec, secrets, err := s.plugin.GetScanJobSpec(spec, credentials)
 	if err != nil {
 		return nil, nil, err
 	}
 
-	templateSpec.Tolerations = append(templateSpec.Tolerations, getScanJobTolerations(starboardConfigData)...)
+	scanJobTolerations, err := starboardConfigData.GetScanJobTolerations()
+	if err != nil {
+		return nil, nil, err
+	}
+	templateSpec.Tolerations = append(templateSpec.Tolerations, scanJobTolerations...)
 
 	templateSpec.ServiceAccountName = starboard.ServiceAccountName
